Get both file and git sources via /source/{sourceId}
