--- tests/Functional/Application/FileSource/GetTest.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Tests\Functional\Application\FileSource;
-
-use App\Tests\Application\FileSource\AbstractGetTest;
-use App\Tests\Functional\GetClientAdapterTrait;
-
-class GetTest extends AbstractGetTest
-{
-    use GetClientAdapterTrait;
-}
