+++ tests/Functional/Controller/Source/SourceControllerTest.php
@@ -1,0 +1,70 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Tests\Functional\Controller\Source;
+
+use App\Tests\Application\AbstractApplicationTestCase;
+use App\Tests\DataProvider\InvalidResponseModelDataProviderCreatorTrait;
+use App\Tests\DataProvider\ServiceHttpFailureDataProviderCreatorTrait;
+use App\Tests\Functional\Controller\AssertJsonResponseTrait;
+use App\Tests\Functional\GetClientAdapterTrait;
+use SmartAssert\SourcesClient\GitSourceClientInterface;
+use SmartAssert\UsersClient\ClientInterface as UsersClient;
+use SmartAssert\UsersClient\Model\Token;
+use Symfony\Component\Uid\Ulid;
+
+class SourceControllerTest extends AbstractApplicationTestCase
+{
+    use GetClientAdapterTrait;
+    use ServiceHttpFailureDataProviderCreatorTrait;
+    use InvalidResponseModelDataProviderCreatorTrait;
+    use AssertJsonResponseTrait;
+
+    /**
+     * @dataProvider sourcesClientExceptionDataProvider
+     *
+     * @param array<mixed> $expectedData
+     */
+    public function testGetHandlesException(
+        \Exception $exception,
+        int $expectedStatusCode,
+        array $expectedData
+    ): void {
+        $apiKey = md5((string) rand());
+        $apiToken = md5((string) rand());
+        $sourceId = (string) new Ulid();
+
+        $usersClient = \Mockery::mock(UsersClient::class);
+        $usersClient
+            ->shouldReceive('createApiToken')
+            ->with($apiKey)
+            ->andReturn(new Token($apiToken))
+        ;
+
+        $gitSourceClient = \Mockery::mock(GitSourceClientInterface::class);
+        $gitSourceClient
+            ->shouldReceive('get')
+            ->with($apiToken, $sourceId)
+            ->andThrow($exception)
+        ;
+
+        self::getContainer()->set(UsersClient::class, $usersClient);
+        self::getContainer()->set(GitSourceClientInterface::class, $gitSourceClient);
+
+        $response = $this->applicationClient->makeGetSourceRequest($apiKey, $sourceId);
+
+        $this->assertJsonResponse($response, $expectedStatusCode, $expectedData);
+    }
+
+    /**
+     * @return array<mixed>
+     */
+    public function sourcesClientExceptionDataProvider(): array
+    {
+        return array_merge(
+            $this->serviceHttpFailureDataProviderCreator('sources'),
+            $this->invalidResponseModelDataProviderCreator('sources'),
+        );
+    }
+}
