--- tests/Integration/FileSource/GetTest.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Tests\Integration\FileSource;
-
-use App\Tests\Application\FileSource\AbstractGetTest;
-use App\Tests\Integration\GetClientAdapterTrait;
-
-class GetTest extends AbstractGetTest
-{
-    use GetClientAdapterTrait;
-}
